% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{add_user_to_group}
\alias{add_user_to_group}
\title{Adds the specified user to the specified group}
\usage{
add_user_to_group(GroupName, UserName)
}
\arguments{
\item{GroupName}{The name of the group to update.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{UserName}{The name of the user to add.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Adds the specified user to the specified group.
}
\examples{
# The following command adds an IAM user named Bob to the IAM group named
# Admins:
add_user_to_group(
  GroupName = "Admins",
  UserName = "Bob"
)

}
