% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{attach_group_policy}
\alias{attach_group_policy}
\title{Attaches the specified managed policy to the specified IAM group}
\usage{
attach_group_policy(GroupName, PolicyArn)
}
\arguments{
\item{GroupName}{The name (friendly name, not ARN) of the group to attach the policy to.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{PolicyArn}{The Amazon Resource Name (ARN) of the IAM policy you want to attach.

For more information about ARNs, see \href{http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs) and AWS Service Namespaces} in the \emph{AWS General Reference}.}
}
\description{
Attaches the specified managed policy to the specified IAM group.
}
\details{
You use this API to attach a managed policy to a group. To embed an inline policy in a group, use PutGroupPolicy.

For more information about policies, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies-managed-vs-inline.html}{Managed Policies and Inline Policies} in the \emph{IAM User Guide}.
}
\examples{
# The following command attaches the AWS managed policy named
# ReadOnlyAccess to the IAM group named Finance.
attach_group_policy(
  GroupName = "Finance",
  PolicyArn = "arn:aws:iam::aws:policy/ReadOnlyAccess"
)

}
