% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{create_login_profile}
\alias{create_login_profile}
\title{Creates a password for the specified user, giving the user the ability to access AWS services through the AWS Management Console}
\usage{
create_login_profile(
  UserName = "string",
  Password = "string",
  PasswordResetRequired = TRUE|FALSE
)
}
\arguments{
\item{UserName}{[required] The name of the IAM user to create a password for. The user must already exist.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{Password}{[required] The new password for the user.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the AWS Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.}

\item{PasswordResetRequired}{Specifies whether the user is required to set a new password on next sign-in.}
}
\description{
Creates a password for the specified user, giving the user the ability to access AWS services through the AWS Management Console. For more information about managing passwords, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/Using_ManagingLogins.html}{Managing Passwords} in the \emph{IAM User Guide}.
}
\examples{
# The following command changes IAM user Bob's password and sets the flag
# that required Bob to change the password the next time he signs in.
create_login_profile(
  Password = "h]6EszR}vJ*m",
  PasswordResetRequired = TRUE,
  UserName = "Bob"
)

}
