% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{create_service_linked_role}
\alias{create_service_linked_role}
\title{Creates an IAM role that is linked to a specific AWS service}
\usage{
create_service_linked_role(AWSServiceName, Description = NULL,
  CustomSuffix = NULL)
}
\arguments{
\item{AWSServiceName}{The AWS service to which this role is attached. You use a string similar to a URL but without the http:// in front. For example: \code{elasticbeanstalk.amazonaws.com}}

\item{Description}{The description of the role.}

\item{CustomSuffix}{A string that you provide, which is combined with the service name to form the complete role name. If you make multiple requests for the same service, then you must supply a different \code{CustomSuffix} for each request. Otherwise the request fails with a duplicate role name error. For example, you could add \code{-1} or \code{-debug} to the suffix.}
}
\description{
Creates an IAM role that is linked to a specific AWS service. The service controls the attached policies and when the role can be deleted. This helps ensure that the service is not broken by an unexpectedly changed or deleted role, which could put your AWS resources into an unknown state. Allowing the service to control the role helps improve service stability and proper cleanup when a service and its role are no longer needed.
}
\details{
The name of the role is generated by combining the string that you specify for the \code{AWSServiceName} parameter with the string that you specify for the \code{CustomSuffix} parameter. The resulting name must be unique in your account or the request fails.

To attach a policy to this service-linked role, you must make the request using the AWS service that depends on this role.
}
\examples{

}
