% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{create_user}
\alias{create_user}
\title{Creates a new IAM user for your AWS account}
\usage{
create_user(UserName, Path = NULL, PermissionsBoundary = NULL)
}
\arguments{
\item{UserName}{The name of the user to create.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-. User names are not distinguished by case. For example, you cannot create users named both "TESTUSER" and "testuser".}

\item{Path}{The path for the user name. For more information about paths, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/Using_Identifiers.html}{IAM Identifiers} in the \emph{IAM User Guide}.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.}

\item{PermissionsBoundary}{The ARN of the policy that is used to set the permissions boundary for the user.}
}
\description{
Creates a new IAM user for your AWS account.
}
\details{
For information about limitations on the number of IAM users you can create, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/LimitationsOnEntities.html}{Limitations on IAM Entities} in the \emph{IAM User Guide}.
}
\examples{
# The following create-user command creates an IAM user named Bob in the
# current account.
create_user(
  UserName = "Bob"
)

}
