% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{delete_group}
\alias{delete_group}
\title{Deletes the specified IAM group}
\usage{
delete_group(
  GroupName = "string"
)
}
\arguments{
\item{GroupName}{[required] The name of the IAM group to delete.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Deletes the specified IAM group. The group must not contain any users or have any attached policies.
}
\examples{

}
