% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{delete_instance_profile}
\alias{delete_instance_profile}
\title{Deletes the specified instance profile}
\usage{
delete_instance_profile(
  InstanceProfileName = "string"
)
}
\arguments{
\item{InstanceProfileName}{[required] The name of the instance profile to delete.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Deletes the specified instance profile. The instance profile must not have an associated role.
}
\details{
Make sure that you do not have any Amazon EC2 instances running with the instance profile you are about to delete. Deleting a role or instance profile that is associated with a running instance will break any applications running on the instance.

For more information about instance profiles, go to \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/AboutInstanceProfiles.html}{About Instance Profiles}.
}
\examples{
# The following command deletes the instance profile named
# ExampleInstanceProfile
delete_instance_profile(
  InstanceProfileName = "ExampleInstanceProfile"
)

}
