% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{delete_role_permissions_boundary}
\alias{delete_role_permissions_boundary}
\title{Deletes the permissions boundary for the specified IAM role}
\usage{
delete_role_permissions_boundary(
  RoleName = "string"
)
}
\arguments{
\item{RoleName}{[required] The name (friendly name, not ARN) of the IAM role from which you want to remove the permissions boundary.}
}
\description{
Deletes the permissions boundary for the specified IAM role.
}
\details{
Deleting the permissions boundary for a role might increase its permissions by allowing anyone who assumes the role to perform all the actions granted in its permissions policies.
}
\examples{

}
