% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{delete_user_permissions_boundary}
\alias{delete_user_permissions_boundary}
\title{Deletes the permissions boundary for the specified IAM user}
\usage{
delete_user_permissions_boundary(UserName)
}
\arguments{
\item{UserName}{The name (friendly name, not ARN) of the IAM user from which you want to remove the permissions boundary.}
}
\description{
Deletes the permissions boundary for the specified IAM user.
}
\details{
Deleting the permissions boundary for a user might increase its permissions by allowing the user to perform all the actions granted in its permissions policies.
}
\examples{

}
