% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{delete_user_policy}
\alias{delete_user_policy}
\title{Deletes the specified inline policy that is embedded in the specified IAM user}
\usage{
delete_user_policy(
  UserName = "string",
  PolicyName = "string"
)
}
\arguments{
\item{UserName}{[required] The name (friendly name, not ARN) identifying the user that the policy is embedded in.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{PolicyName}{[required] The name identifying the policy document to delete.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Deletes the specified inline policy that is embedded in the specified IAM user.
}
\details{
A user can also have managed policies attached to it. To detach a managed policy from a user, use DetachUserPolicy. For more information about policies, refer to \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies-managed-vs-inline.html}{Managed Policies and Inline Policies} in the \emph{IAM User Guide}.
}
\examples{
# The following delete-user-policy command removes the specified policy
# from the IAM user named Juan:
delete_user_policy(
  PolicyName = "ExamplePolicy",
  UserName = "Juan"
)

}
