% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{generate_credential_report}
\alias{generate_credential_report}
\title{Generates a credential report for the AWS account}
\usage{
generate_credential_report()
}
\description{
Generates a credential report for the AWS account. For more information about the credential report, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/credential-reports.html}{Getting Credential Reports} in the \emph{IAM User Guide}.
}
