% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_access_key_last_used}
\alias{get_access_key_last_used}
\title{Retrieves information about when the specified access key was last used}
\usage{
get_access_key_last_used(AccessKeyId)
}
\arguments{
\item{AccessKeyId}{The identifier of an access key.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters that can consist of any upper or lowercased letter or digit.}
}
\description{
Retrieves information about when the specified access key was last used. The information includes the date and time of last use, along with the AWS service and region that were specified in the last request made with that key.
}
\examples{

}
