% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_account_password_policy}
\alias{get_account_password_policy}
\title{Retrieves the password policy for the AWS account}
\usage{
get_account_password_policy()
}
\description{
Retrieves the password policy for the AWS account. For more information about using a password policy, go to \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/Using_ManagingPasswordPolicies.html}{Managing an IAM Password Policy}.
}
\examples{
# The following command displays details about the password policy for the
# current AWS account.
get_account_password_policy()

}
