% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_account_summary}
\alias{get_account_summary}
\title{Retrieves information about IAM entity usage and IAM quotas in the AWS account}
\usage{
get_account_summary()
}
\description{
Retrieves information about IAM entity usage and IAM quotas in the AWS account.
}
\details{
For information about limitations on IAM entities, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/LimitationsOnEntities.html}{Limitations on IAM Entities} in the \emph{IAM User Guide}.
}
\examples{
# The following command returns information about the IAM entity quotas
# and usage in the current AWS account.
get_account_summary()

}
