% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_context_keys_for_custom_policy}
\alias{get_context_keys_for_custom_policy}
\title{Gets a list of all of the context keys referenced in the input policies}
\usage{
get_context_keys_for_custom_policy(
  PolicyInputList = list(
    "string"
  )
)
}
\arguments{
\item{PolicyInputList}{[required] A list of policies for which you want the list of context keys referenced in those policies. Each document is specified as a string containing the complete, valid JSON text of an IAM policy.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} used to validate this parameter is a string of characters consisting of the following:
\itemize{
\item Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range
\item The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)
\item The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)
}}
}
\description{
Gets a list of all of the context keys referenced in the input policies. The policies are supplied as a list of one or more strings. To get the context keys from policies associated with an IAM user, group, or role, use GetContextKeysForPrincipalPolicy.
}
\details{
Context keys are variables maintained by AWS and its services that provide details about the context of an API query request. Context keys can be evaluated by testing against a value specified in an IAM policy. Use \code{GetContextKeysForCustomPolicy} to understand what key names and values you must supply when you call SimulateCustomPolicy. Note that all parameters are shown in unencoded form here for clarity but must be URL encoded to be included as a part of a real HTML request.
}
\examples{

}
