% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_group_policy}
\alias{get_group_policy}
\title{Retrieves the specified inline policy document that is embedded in the specified IAM group}
\usage{
get_group_policy(GroupName, PolicyName)
}
\arguments{
\item{GroupName}{The name of the group the policy is associated with.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{PolicyName}{The name of the policy document to get.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Retrieves the specified inline policy document that is embedded in the specified IAM group.
}
\details{
Policies returned by this API are URL-encoded compliant with \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}. You can use a URL decoding method to convert the policy back to plain JSON text. For example, if you use Java, you can use the \code{decode} method of the \code{java.net.URLDecoder} utility class in the Java SDK. Other languages and SDKs provide similar functionality.

An IAM group can also have managed policies attached to it. To retrieve a managed policy document that is attached to a group, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies-managed-vs-inline.html}{Managed Policies and Inline Policies} in the \emph{IAM User Guide}.
}
\examples{

}
