% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_saml_provider}
\alias{get_saml_provider}
\title{Returns the SAML provider metadocument that was uploaded when the IAM SAML provider resource object was created or updated}
\usage{
get_saml_provider(SAMLProviderArn)
}
\arguments{
\item{SAMLProviderArn}{[required] The Amazon Resource Name (ARN) of the SAML provider resource object in IAM to get information about.

For more information about ARNs, see \href{http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs) and AWS Service Namespaces} in the \emph{AWS General Reference}.}
}
\description{
Returns the SAML provider metadocument that was uploaded when the IAM SAML provider resource object was created or updated.
}
\details{
This operation requires \href{http://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4}.
}
\section{Accepted Parameters}{
\preformatted{get_saml_provider(
  SAMLProviderArn = "string"
)
}
}

