% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{get_server_certificate}
\alias{get_server_certificate}
\title{Retrieves information about the specified server certificate stored in IAM}
\usage{
get_server_certificate(ServerCertificateName)
}
\arguments{
\item{ServerCertificateName}{[required] The name of the server certificate you want to retrieve information about.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Retrieves information about the specified server certificate stored in IAM.
}
\details{
For more information about working with server certificates, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_server-certs.html}{Working with Server Certificates} in the \emph{IAM User Guide}. This topic includes a list of AWS services that can use the server certificates that you manage with IAM.
}
\section{Accepted Parameters}{
\preformatted{get_server_certificate(
  ServerCertificateName = "string"
)
}
}

