% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{list_account_aliases}
\alias{list_account_aliases}
\title{Lists the account alias associated with the AWS account (Note: you can have only one)}
\usage{
list_account_aliases(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the \code{Marker} element in the response that you received to indicate where the next call should start.}

\item{MaxItems}{Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the \code{IsTruncated} response element is \code{true}.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the \code{IsTruncated} response element returns \code{true}, and \code{Marker} contains a value to include in the subsequent call that tells the service where to continue from.}
}
\description{
Lists the account alias associated with the AWS account (Note: you can have only one). For information about using an AWS account alias, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/AccountAlias.html}{Using an Alias for Your AWS Account ID} in the \emph{IAM User Guide}.
}
\section{Accepted Parameters}{
\preformatted{list_account_aliases(
  Marker = "string",
  MaxItems = 123
)
}
}

\examples{
# The following command lists the aliases for the current account.
\donttest{list_account_aliases()}

}
