% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{list_open_id_connect_providers}
\alias{list_open_id_connect_providers}
\title{Lists information about the IAM OpenID Connect (OIDC) provider resource objects defined in the AWS account}
\usage{
list_open_id_connect_providers()
}
\description{
Lists information about the IAM OpenID Connect (OIDC) provider resource objects defined in the AWS account.
}
\examples{

}
