% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{list_saml_providers}
\alias{list_saml_providers}
\title{Lists the SAML provider resource objects defined in IAM in the account}
\usage{
list_saml_providers()
}
\description{
Lists the SAML provider resource objects defined in IAM in the account.
}
\details{
This operation requires \href{http://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4}.
}
\examples{

}
