% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{list_user_policies}
\alias{list_user_policies}
\title{Lists the names of the inline policies embedded in the specified IAM user}
\usage{
list_user_policies(UserName, Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{UserName}{[required] The name of the user to list policies for.

This parameter allows (per its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{Marker}{Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the \code{Marker} element in the response that you received to indicate where the next call should start.}

\item{MaxItems}{(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the \code{IsTruncated} response element is \code{true}.

If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the \code{IsTruncated} response element returns \code{true} and \code{Marker} contains a value to include in the subsequent call that tells the service where to continue from.}
}
\description{
Lists the names of the inline policies embedded in the specified IAM user.
}
\details{
An IAM user can also have managed policies attached to it. To list the managed policies that are attached to a user, use ListAttachedUserPolicies. For more information about policies, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies-managed-vs-inline.html}{Managed Policies and Inline Policies} in the \emph{IAM User Guide}.

You can paginate the results using the \code{MaxItems} and \code{Marker} parameters. If there are no inline policies embedded with the specified user, the operation returns an empty list.
}
\section{Accepted Parameters}{
\preformatted{list_user_policies(
  UserName = "string",
  Marker = "string",
  MaxItems = 123
)
}
}

