% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_package.R
\docType{package}
\name{paws.iam-package}
\alias{paws.iam}
\alias{paws.iam-package}
\title{paws.iam: AWS Identity and Access Management}
\description{
AWS Identity and Access Management (IAM) is a web service
that you can use to manage users and user permissions under your AWS
account. This guide provides descriptions of IAM actions that you can
call programmatically. For general information about IAM, see [AWS
Identity and Access Management (IAM)](http://aws.amazon.com/iam/). For
the user guide for IAM, see [Using
IAM](http://docs.aws.amazon.com/IAM/latest/UserGuide/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
