% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{remove_user_from_group}
\alias{remove_user_from_group}
\title{Removes the specified user from the specified group}
\usage{
remove_user_from_group(
  GroupName = "string",
  UserName = "string"
)
}
\arguments{
\item{GroupName}{[required] The name of the group to update.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{UserName}{[required] The name of the user to remove.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}
}
\description{
Removes the specified user from the specified group.
}
\examples{
# The following command removes the user named Bob from the IAM group
# named Admins.
remove_user_from_group(
  GroupName = "Admins",
  UserName = "Bob"
)

}
