% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{untag_role}
\alias{untag_role}
\title{Removes the specified tags from the role}
\usage{
untag_role(RoleName, TagKeys)
}
\arguments{
\item{RoleName}{[required] The name of the IAM role from which you want to remove tags.

This parameter accepts (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters that consist of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: \_+=,.@-}

\item{TagKeys}{[required] A list of key names as a simple array of strings. The tags with matching keys are removed from the specified role.}
}
\description{
Removes the specified tags from the role. For more information about tagging, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM Identities} in the \emph{IAM User Guide}.
}
\section{Accepted Parameters}{
\preformatted{untag_role(
  RoleName = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\examples{
# The following example shows how to remove a tag with the key 'Dept' from
# a role named 'taggedrole'.
\donttest{untag_role(
  RoleName = "taggedrole",
  TagKeys = list(
    "Dept"
  )
)}

}
