% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iam_operations.R
\name{update_role_description}
\alias{update_role_description}
\title{Use instead}
\usage{
update_role_description(RoleName, Description)
}
\arguments{
\item{RoleName}{The name of the role that you want to modify.}

\item{Description}{The new description that you want to apply to the specified role.}
}
\description{
Use instead.
}
\details{
Modifies only the description of a role. This operation performs the same function as the \code{Description} parameter in the \code{UpdateRole} operation.
}
\examples{

}
